Finished user control. All ednpoint chekced.

+++ server/Services/CanvasesService.cs
@@ -17,10 +17,13 @@ namespace server.Services
             _imagesService = imagesService;
         }
 
-        public async Task<List<CanvasListItem>> GetCanvasesForUserAsync()
+        public async Task<List<CanvasListItem>> GetCanvasesForUserAsync(string uid)
         {
             CollectionReference canvasesRef = _firestoreDb.Collection("canvases");
-            QuerySnapshot snapshot = await canvasesRef.GetSnapshotAsync();
+
+            Query userCanvases = canvasesRef.WhereEqualTo("UserId", uid);
+
+            QuerySnapshot snapshot = await userCanvases.GetSnapshotAsync();
 
             List<CanvasListItem> result = snapshot.Documents.Select(doc => new CanvasListItem
             {
@@ -34,18 +37,18 @@ namespace server.Services
             return result;
         }
 
-        public async Task<Canvas> CreateCanvasAsync(CreateCanvasRequest request)
+        public async Task<Canvas> CreateCanvasAsync(CreateCanvasRequest request, string uid)
         {
             DocumentReference addedDocRef = _firestoreDb.Collection("canvases").Document();
 
-            Canvas newCanvas = Canvas.CreateNew(addedDocRef.Id, "ADMIN", request.Name ?? "Untitled Canvas");
+            Canvas newCanvas = Canvas.CreateNew(addedDocRef.Id, uid, request.Name ?? "Untitled Canvas");
 
             await addedDocRef.SetAsync(newCanvas);
 
             return newCanvas;
         }
 
-        public async Task DeleteCanvasAsync(string canvasId)
+        public async Task DeleteCanvasAsync(string canvasId, string uid)
         {
             // First get the canvas to retrieve the userId
             DocumentReference canvasRef = _firestoreDb.Collection("canvases").Document(canvasId);
@@ -53,16 +56,22 @@ namespace server.Services
             
             if (!snapshot.Exists)
             {
-                throw new Exception("Canvas not found");
+                throw new KeyNotFoundException("Canvas not found.");
             }
             
             var canvas = snapshot.ConvertTo<Canvas>();
-            string userId = canvas.UserId;
-            
+
+            string canvasUserId = canvas.UserId;
+
+            if (canvas.UserId != uid)
+            {
+                throw new UnauthorizedAccessException("You do not have permission to delete this canvas.");
+            }
+
             try
             {
                 // Delete all images associated with this canvas
-                await _imagesService.DeleteAllCanvasImagesAsync(userId, canvasId);
+                await _imagesService.DeleteAllCanvasImagesAsync(canvasUserId, canvasId);
                 
                 await canvasRef.DeleteAsync();
             }
@@ -72,7 +81,7 @@ namespace server.Services
             }
         }
 
-        public async Task<Canvas> GetCanvasAsync(string canvasId)
+        public async Task<Canvas> GetCanvasAsync(string canvasId, string uid)
         {
             DocumentReference canvasRef = _firestoreDb.Collection("canvases").Document(canvasId);
             DocumentSnapshot snapshot = await canvasRef.GetSnapshotAsync();
@@ -82,10 +91,17 @@ namespace server.Services
                 throw new Exception("Canvas not found");
             }
 
-            return snapshot.ConvertTo<Canvas>();
+            var canvas = snapshot.ConvertTo<Canvas>();
+
+            if (canvas.UserId != uid)
+            {
+                throw new UnauthorizedAccessException("You do not have permission to see this canvas.");
+            }
+
+            return canvas;
         }
 
-        public async Task<Canvas> UpdateCanvasAsync(string canvasId, Canvas canvas)
+        public async Task<Canvas> UpdateCanvasAsync(string canvasId, UpdateCanvasRequest request, string uid)
         {
             DocumentReference canvasRef = _firestoreDb.Collection("canvases").Document(canvasId);
             DocumentSnapshot snapshot = await canvasRef.GetSnapshotAsync();
@@ -95,11 +111,17 @@ namespace server.Services
                 throw new Exception("Canvas not found");
             }
 
-            // Update timestamp before saving
-            canvas.UpdatedAt = DateTime.UtcNow;
+            var canvas = snapshot.ConvertTo<Canvas>();
+
+            if (canvas.UserId != uid)
+            {
+                throw new UnauthorizedAccessException("You do not have permission to see this canvas.");
+            }
+
+            Canvas updatedCanvas = Canvas.Update(canvas, request.Data, request.Elements);
             
-            await canvasRef.SetAsync(canvas);
-            return canvas;
+            await canvasRef.SetAsync(updatedCanvas);
+            return updatedCanvas;
         }
     }
 }
